remove biaffine from current models and refactor as an individual module
