+++ src/transformer_biaffine_ner/data_utils.py
@@ -1,3 +1,8 @@
 # coding=utf-8
 # Created by bugface (https://github.com/bugface)
 # First created at 12/31/21
+
+
+class TransformerNerBiaffineDataProcessor(object):
+    def __init__(self):
+        pass
